+++ SasquatchMac/SasquatchMac/AppCenterDelegate.swift
@@ -15,7 +15,9 @@ import AppCenter
   func appSecret() -> String
   func logUrl() -> String
   func isDebuggerAttached() -> Bool
+  func startAnalyticsFromLibrary()
   func setUserId(_ userId: String?)
+  func setCountryCode(_ countryCode: String?)
 
   // Modules section.
   func isAnalyticsEnabled() -> Bool
@@ -28,8 +30,13 @@ import AppCenter
   // MSAnalytics section.
   func trackEvent(_ eventName: String)
   func trackEvent(_ eventName: String, withProperties: Dictionary<String, String>)
+  func trackEvent(_ eventName: String, withProperties: Dictionary<String, String>, flags: MSFlags)
+  func trackEvent(_ eventName: String, withTypedProperties: MSEventProperties)
+  func trackEvent(_ eventName: String, withTypedProperties: MSEventProperties?, flags: MSFlags)
   func trackPage(_ pageName: String)
   func trackPage(_ pageName: String, withProperties: Dictionary<String, String>)
+  func resume()
+  func pause()
 
   // MSCrashes section.
   func hasCrashedInLastSession() -> Bool