+++ SasquatchMac/SasquatchMac/ViewControllers/AnalyticsViewController.swift
@@ -3,9 +3,46 @@ import Cocoa
 // FIXME: trackPage has been hidden in MSAnalytics temporarily. Use internal until the feature comes back.
 class AnalyticsViewController : NSViewController, NSTableViewDataSource, NSTableViewDelegate {
 
-  private enum CellIdentifiers {
-    static let keyCellId = "keyCellId"
-    static let valueCellId = "valueCellId"
+  class EventProperty : NSObject {
+    var key: String = ""
+    var type: String = EventPropertyType.string.rawValue
+    var string: String = ""
+    var double: NSNumber = 0
+    var long: NSNumber = 0
+    var boolean: Bool = false
+    var dateTime: Date = Date.init()
+  }
+
+  enum EventPropertyType : String {
+    case string = "String"
+    case double = "Double"
+    case long = "Long"
+    case boolean = "Boolean"
+    case dateTime = "DateTime"
+
+    static let allValues = [string, double, long, boolean, dateTime]
+  }
+
+  enum Priority: String {
+    case defaultType = "Default"
+    case normal = "Normal"
+    case critical = "Critical"
+    case invalid = "Invalid"
+
+    var flags: MSFlags {
+      switch self {
+      case .normal:
+        return [.persistenceNormal]
+      case .critical:
+        return [.persistenceCritical]
+      case .invalid:
+        return MSFlags.init(rawValue: 42)
+      default:
+        return []
+      }
+    }
+
+    static let allValues = [defaultType, normal, critical, invalid]
   }
 
   var appCenter: AppCenterDelegate = AppCenterProvider.shared().appCenter!
@@ -13,18 +50,23 @@ class AnalyticsViewController : NSViewController, NSTableViewDataSource, NSTable
   @IBOutlet weak var name: NSTextField!
   @IBOutlet var setEnabledButton : NSButton?
   @IBOutlet var table : NSTableView?
+  @IBOutlet weak var pause: NSButton!
+  @IBOutlet weak var resume: NSButton!
+  @IBOutlet weak var priorityValue: NSComboBox!
+  @IBOutlet weak var countLabel: NSTextField!
+  @IBOutlet weak var countSlider: NSSlider!
+  @IBOutlet var arrayController: NSArrayController!
 
-  private var properties : [String : String] = [String : String]()
   private var textBeforeEditing : String = ""
   private var totalPropsCounter : Int = 0
+  private var priority = Priority.defaultType
+  dynamic var eventProperties = [EventProperty]()
 
   override func viewDidLoad() {
     super.viewDidLoad()
     setEnabledButton?.state = appCenter.isAnalyticsEnabled() ? 1 : 0
     table?.delegate = self
-    table?.dataSource = self
-    NotificationCenter.default.addObserver(self, selector: #selector(self.editingDidBegin), name: .NSControlTextDidBeginEditing, object: nil)
-    NotificationCenter.default.addObserver(self, selector: #selector(self.editingDidEnd), name: .NSControlTextDidEndEditing, object: nil)
+    self.countLabel.stringValue = "Count: \(Int(countSlider.intValue))"
   }
 
   override func viewWillAppear() {
@@ -37,113 +79,167 @@ class AnalyticsViewController : NSViewController, NSTableViewDataSource, NSTable
   }
 
   @IBAction func trackEvent(_ : AnyObject) {
-    appCenter.trackEvent(name.stringValue, withProperties: properties)
+    let eventProperties = eventPropertiesSet()
+    let eventName = name.stringValue
+    for _ in 0..<Int(countSlider.intValue) {
+      if let properties = eventProperties as? MSEventProperties {
+        if priority != .defaultType {
+          appCenter.trackEvent(eventName, withTypedProperties: properties, flags: priority.flags)
+        } else {
+          appCenter.trackEvent(eventName, withTypedProperties: properties)
+        }
+      } else if let dictionary = eventProperties as? [String: String] {
+        if priority != .defaultType {
+          appCenter.trackEvent(eventName, withProperties: dictionary, flags: priority.flags)
+        } else {
+          appCenter.trackEvent(eventName, withProperties: dictionary)
+        }
+      } else {
+        if priority != .defaultType {
+          appCenter.trackEvent(eventName, withTypedProperties: nil, flags: priority.flags)
+        } else {
+          appCenter.trackEvent(eventName)
+        }
+      }
+      for targetToken in TransmissionTargets.shared.transmissionTargets.keys {
+        if TransmissionTargets.shared.targetShouldSendAnalyticsEvents(targetToken: targetToken) {
+          let target = TransmissionTargets.shared.transmissionTargets[targetToken]!
+          if let properties = eventProperties as? MSEventProperties {
+            if priority != .defaultType {
+              target.trackEvent(eventName, withProperties: properties, flags: priority.flags)
+            } else {
+              target.trackEvent(eventName, withProperties: properties)
+            }
+          } else if let dictionary = eventProperties as? [String: String] {
+            if priority != .defaultType {
+              target.trackEvent(eventName, withProperties: dictionary, flags: priority.flags)
+            } else {
+              target.trackEvent(eventName, withProperties: dictionary)
+            }
+          } else {
+            if priority != .defaultType {
+              target.trackEvent(eventName, withProperties: [:], flags: priority.flags)
+            } else {
+              target.trackEvent(eventName)
+            }
+          }
+        }
+      }
+    }
+  }
+
+  @IBAction func resume(_ sender: NSButton) {
+    appCenter.resume()
+  }
+
+  @IBAction func pause(_ sender: NSButton) {
+    appCenter.pause()
+  }
+
+  @IBAction func countChanged(_ sender: Any) {
+    self.countLabel.stringValue = "Count: \(Int(countSlider.intValue))"
+  }
+
+  @IBAction func priorityChanged(_ sender: NSComboBox) {
+    self.priority = Priority(rawValue: self.priorityValue.stringValue)!
   }
 
   @IBAction func trackPage(_ : AnyObject) {
-    NSLog("trackPageWithProperties: %d", properties.count)
+    NSLog("trackPageWithProperties: %d", eventProperties.count)
   }
 
   @IBAction func addProperty(_ : AnyObject) {
-    let newKey = String(format:"key%d",totalPropsCounter)
-    let newValue = String(format:"value%d",totalPropsCounter)
-
-    self.properties.updateValue(newValue, forKey: newKey)
-    table?.reloadData()
-
-    totalPropsCounter+=1
+    let property = EventProperty()
+    let eventProperties = arrayController.content as! [EventProperty]
+    let count = eventProperties.count
+    property.key = "key\(count)"
+    property.string = "value\(count)"
+    property.addObserver(self, forKeyPath: #keyPath(EventProperty.type), options: .new, context: nil)
+    arrayController.addObject(property)
   }
 
   @IBAction func deleteProperty(_ : AnyObject) {
-    if properties.isEmpty {
-      return
+    if let selectedProperty = arrayController.selectedObjects.first as? EventProperty {
+      arrayController.removeObject(selectedProperty)
+      selectedProperty.removeObserver(self, forKeyPath: #keyPath(EventProperty.type), context: nil)
     }
-    guard let `table` = table else {
-      return
-    }
-    if (table.selectedRow < 0) {
-      _ = properties.popFirst()
-    } else {
-      let key : String = Array(properties.keys)[table.selectedRow]
-      _ = properties.removeValue(forKey: key)
-    }
-    table.reloadData()
   }
 
   @IBAction func setEnabled(sender : NSButton) {
     appCenter.setAnalyticsEnabled(sender.state == 1)
     sender.state = appCenter.isAnalyticsEnabled() ? 1 : 0
   }
-
-  //MARK: Table view source delegate
-
-  func numberOfRows(in tableView: NSTableView) -> Int {
-    return properties.count
-  }
-
-  //MARK: Table view delegate
   
   func tableView(_ tableView: NSTableView, viewFor tableColumn: NSTableColumn?, row: Int) -> NSView? {
-    guard let `tableColumn` = tableColumn else {
+    guard let identifier = tableColumn?.identifier else {
       return nil
     }
-
-    var cellValue : String = ""
-    var cellId : String = ""
-    let key : String = Array(properties.keys)[row]
-
-    if (tableColumn == tableView.tableColumns[0]) {
-      cellValue = key
-      cellId = CellIdentifiers.keyCellId
-    } else if (tableColumn == tableView.tableColumns[1]) {
-      cellValue = properties[key]!
-      cellId = CellIdentifiers.valueCellId
+    let view = tableView.make(withIdentifier: identifier, owner: self)
+    if (identifier == "value") {
+      updateValue(property: eventProperties[row], cell: view as! NSTableCellView)
     }
-
-    if let cell = tableView.make(withIdentifier: cellId, owner: nil) as? NSTableCellView {
-      cell.textField?.stringValue = cellValue
-      cell.textField?.isEditable = true
-      return cell
-    }
-
-    return nil
+    return view
   }
 
-  //MARK: Text field events
-
-  func editingDidBegin(notification : NSNotification) {
-    guard let textField = notification.object as? NSTextField else {
+  override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey : Any]?, context: UnsafeMutableRawPointer?) {
+    guard let property = object as? EventProperty else {
       return
     }
-    textBeforeEditing = textField.stringValue
+    guard let row = eventProperties.index(of: property) else {
+      return
+    }
+    let column = table?.column(withIdentifier: "value")
+    guard let cell = table?.view(atColumn: column!, row: row, makeIfNecessary: false) as? NSTableCellView else {
+      return
+    }
+    updateValue(property: property, cell: cell)
   }
 
-  func editingDidEnd(notification : NSNotification) {
-    guard let textField = notification.object as? NSTextField else {
+  func updateValue(property: EventProperty, cell: NSTableCellView) {
+    cell.isHidden = false
+    for subview in cell.subviews {
+      subview.isHidden = true
+    }
+    guard let type = EventPropertyType(rawValue: property.type) else {
       return
     }
+    if let view = cell.viewWithTag(EventPropertyType.allValues.index(of: type)!) {
+      view.isHidden = false
+    } else {
+      cell.isHidden = true
+    }
+  }
 
-    // If key
-    if (properties.keys.contains(textBeforeEditing)) {
-      let oldKey : String = textBeforeEditing
-      let newKey : String = textField.stringValue
-      if let value = properties.removeValue(forKey: oldKey) {
-        properties.updateValue(value, forKey: newKey)
+  func eventPropertiesSet() -> Any? {
+    if eventProperties.count < 1 {
+      return nil
+    }
+    var onlyStrings = true
+    var propertyDictionary = [String: String]()
+    let properties = MSEventProperties()
+    for property in eventProperties {
+      let key = property.key
+      guard let type = EventPropertyType(rawValue: property.type) else {
+        continue
+      }
+      switch type {
+      case .string:
+        properties.setEventProperty(property.string, forKey: key);
+        propertyDictionary[property.key] = property.string
+      case .double:
+        properties.setEventProperty(property.double.doubleValue, forKey: key)
+        onlyStrings = false
+      case .long:
+        properties.setEventProperty(property.long.int64Value, forKey: key)
+        onlyStrings = false
+      case .boolean:
+        properties.setEventProperty(property.boolean, forKey: key)
+        onlyStrings = false
+      case .dateTime:
+        properties.setEventProperty(property.dateTime, forKey: key)
+        onlyStrings = false
       }
     }
-
-    // If value
-    else {
-      guard let row = table?.row(for: textField) else {
-        return
-      }
-      if row < 0 {
-        return
-      }
-      let key : String = Array(properties.keys)[row]
-      properties.updateValue(textField.stringValue, forKey: key)
-    }
-
-    table?.reloadData()
+    return onlyStrings ? propertyDictionary : properties
   }
 }